linkedin_hero/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
WTForms
PyMySQL
cryptography
python-dotenv
Flask-Dance
Authlib
Werkzeug
Jinja2 